28.04.2023: api

+++ parking/manager/views.py
@@ -1,4 +1,5 @@
 import json
+import time
 from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
 from django.views.decorators.csrf import ensure_csrf_cookie
@@ -41,7 +42,7 @@ from rest_framework.permissions import IsAuthenticated
 class HelloWorld(APIView):
     # authentication_classes = [BasicAuthentication]
     # permission_classes = [IsAuthenticated]
-    
+    permission_classes = (permissions.AllowAny,)
     def get(self, request): 
         content = {'message': 'Hello, World!'}
         return Response(content)
@@ -50,9 +51,11 @@ class HelloWorld(APIView):
 
 class AccountApiView(APIView):
     def get(self, request, *args, **kwargs):
-        # id1 = json.loads(request.body)["id"]
         id = self.request.query_params.get('id', None)
-        accounts = Account.objects.filter(id=id)
+        if id is None:
+            accounts = Account.objects.all()
+        else:
+            accounts = Account.objects.filter(id=id)
         serializer = AccountSerializer(accounts, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
@@ -77,26 +80,35 @@ class AccountApiView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self,request):
-        ids = json.loads(request.body)["ids"]    
-        for id in ids:
-            try:
-                account = Account.objects.filter(id=id).first() 
-            except Account.DoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
-            account.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        try:
+            ids = json.loads(request.body)["ids"]
+            if not isinstance(ids, list):
+                return Response({"error": "The ids parameter must be a list."}, status=status.HTTP_400_BAD_REQUEST)
+            for id in ids:
+                try:
+                    account = Account.objects.filter(id=id).first() 
+                except Account.DoesNotExist:
+                    return Response("Account not found",status=status.HTTP_404_NOT_FOUND)
+                account.delete()
+            return Response("Delete successfully",status=status.HTTP_204_NO_CONTENT)
+        except json.JSONDecodeError:
+            return Response({"error": "Invalid JSON data."}, status=status.HTTP_400_BAD_REQUEST)
+        except Exception as e:
+        # Trả về phản hồi khi có lỗi xảy ra
+            return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
     
     
 # API for Log
     
 class LogApiView(APIView):
+    permission_classes = (permissions.AllowAny,)
     def get(self, request):
         logs = Log.objects.all()
         serializer = LogSerializer(logs, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
     def post(self, request):
-        print("HaVE REQUEST")
+        print("HAVE REQUEST")
         serializer = ImageSerializer(data=request.data)
         if serializer.is_valid():
             image = request.data.get('image')
@@ -116,12 +128,10 @@ class LogApiView(APIView):
             duong_dan_luu = folder_url + name_image
             cv2.imwrite(duong_dan_luu, image)
             
-            license_sample = "43A395653"
-            text = ""
+            license_sample = "30E94515"
+            resultDetection = ""
             if image is not None:
-                text = model_AI.mode_AI(image)
-
-                
+                resultDetection = model_AI.mode_AI(image)
                 # cv2.imshow('Image from Cloudinary', image)
                 # cv2.waitKey(0)
                 # cv2.destroyAllWindows()
@@ -129,7 +139,8 @@ class LogApiView(APIView):
                 print('Failed to load image from Cloudinary')
                 
             # websocket
-            print(text)
+            print(resultDetection)
+            
             channel_layer = get_channel_layer()
             async_to_sync(channel_layer.group_send)(
             "test_channel",
@@ -138,40 +149,53 @@ class LogApiView(APIView):
                 "message": cloudinary_url
             }
             )
+            # Thêm log vào database
+            my_data = {'vehicle': resultDetection,
+                       'image_in': cloudinary_url}
+            logSerializer = LogSerializer(data=my_data)
+            vehicle_id = my_data['vehicle']
+            log = Log.objects.filter(vehicle=vehicle_id, time_out__isnull=True).first()
+            # Kiểm tra có tồn tại xe tại bãi chưa
             
-            if text == license_sample:
-                new_filename ="image/license_plate_upload/" + text + " " +  str(timezone.now()) +  ".jpg"
-                uploaded_image = rename(uploaded_image['public_id'], new_filename)
-                serializer.save(name = text,image = uploaded_image['secure_url'])
-                # Thêm log vào database
-                my_data = {'vehicle': text}
-                logSerializer = LogSerializer(data=my_data)
+            if log is not None and log.time_out is None:
+                 return Response("This vehicle is existed",status=400)
+            else:
                 if logSerializer.is_valid():
                     logSerializer.save()
-                    return Response(1,status=201)
-                return Response(logSerializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+            if resultDetection == license_sample:
+                new_filename ="image/license_plate_upload/" + resultDetection + " " +  str(timezone.now()) +  ".jpg"
+                uploaded_image = rename(uploaded_image['public_id'], new_filename)
+                serializer.save(name = resultDetection,image = uploaded_image['secure_url'])
+                return Response(1,status=201)
                 
             else:
-                text = "undefined-vehicle"
-                new_filename ="image/license_plate_upload/" + text + " " +  str(timezone.now()) +  ".jpg"
+                resultDetection = resultDetection + "(unregister)"
+                new_filename ="image/license_plate_upload/" + resultDetection + " " +  str(timezone.now()) +  ".jpg"
                 uploaded_image = rename(uploaded_image['public_id'], new_filename)
-                serializer.save(name = text,image = uploaded_image['secure_url'])
+                serializer.save(name = resultDetection,image = uploaded_image['secure_url'])
                 return Response(0, status=400)   
         else:
             return Response(0, status=400)
         
     
     def put(self, request):
-        text = "30F-123123"
+        # xử lí checkout ....
+        text = "30E94515"
         my_data = {'vehicle': text}
         vehicle = my_data['vehicle']
         try:
-            log = Log.objects.filter(vehicle=vehicle).first()
+            log = Log.objects.filter(vehicle=vehicle, time_out__isnull=True).first()
+            print('before:')
+            print(log)
         except Log.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = LogSerializer(log, data=my_data, partial=True)
         if serializer.is_valid():
             serializer.save()
+            log = Log.objects.filter(vehicle=vehicle).first()
+            print('after: ->>')
+            print(log)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
@@ -187,29 +211,60 @@ class LogApiView(APIView):
     
 # API for Slot
 class SlotApiView(APIView):
+    permission_classes = (permissions.AllowAny,)
     def put(self, request):
+        count = 1
+        var1 = request.PUT.get('var1')
+        var2 = request.PUT.get('var2')
+        var3 = request.PUT.get('var3')
+        var4 = request.PUT.get('var4')
+        varList = [var1,var2,var3,var4]
+        for var in varList:
+           if count==4:
+            try:
+                slot = Slot.objects.filter(id=count).first()
+                count+=1
+            except Slot.DoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+            serializer = SlotSerializer(slot, data=var)
+            if serializer.is_valid():
+                serializer.save()
+        count=1
+        return Response(status=status.HTTP_200_OK)
+    
+class SlotUpdate(APIView):
+    permission_classes = (permissions.AllowAny,)
+    def post(self, request):
         var1 = request.POST.get('var1')
         var2 = request.POST.get('var2')
         var3 = request.POST.get('var3')
         var4 = request.POST.get('var4')
-        try:
-            slot = Slot.objects.filter(id=id).first()
-        except Slot.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        serializer = SlotSerializer(slot, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        varList = [var1,var2,var3,var4]
+        channel_layer = get_channel_layer()
+        async_to_sync(channel_layer.group_send)(
+        "slot_channel",
+        {
+            "type": "slot_update",
+            "message": varList
+        }
+        )
+        num_channels_processed = 0
+        group_name = "slot_channel"
+        while num_channels_processed < len([channel_name for channel_name in channel_layer.channels.keys() if channel_name.startswith(f"group-{group_name}-")]):
+            time.sleep(0.1)
+            num_channels_processed = len([channel_name for channel_name in channel_layer.channels.keys() if channel_name.startswith(f"group-{group_name}-")])
+        return Response(status=status.HTTP_200_OK)
         
 
 # API for Vehicle
     
 class VehicleApiView(APIView):
     def get(self, request):
-        vehicles = Vehicle.objects.all()
-        # id = self.request.query_params.get('id', None)
-        # vehicles = Account.objects.filter(id=id)
+        license_plate = self.request.query_params.get('license_plate', None)
+        if license_plate is None:
+            vehicles = Vehicle.objects.all()
+        else:
+            vehicles = Vehicle.objects.filter(license_plate=license_plate)
         serializer = VehicleSerializer(vehicles, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
@@ -221,9 +276,9 @@ class VehicleApiView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def put(self,request):
-        id = json.loads(request.body)["id"]
+        id = json.loads(request.body)["license_plate"]
         try:
-            account = Vehicle.objects.filter(id=id).first()
+            account = Vehicle.objects.filter(license_plate=id).first()
         except Vehicle.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -234,18 +289,17 @@ class VehicleApiView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self,request):
-        ids = json.loads(request.body)["ids"]
-        for id in ids:
+        license_plates = json.loads(request.body)["license_plates"]
+        for license_plate in license_plates:
             try:
-                account = Vehicle.objects.filter(id=id).first() 
+                account = Vehicle.objects.filter(license_plate=license_plate).first() 
             except Vehicle.DoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+                return Response("Vehicle does not exist",status=status.HTTP_404_NOT_FOUND)
             account.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response("Delete successfully",status=status.HTTP_204_NO_CONTENT)
     
     
 # Get image from esp32
-
 class ImageView(APIView):
     permission_classes = (permissions.AllowAny,)
     def post(self, request):
@@ -312,7 +366,7 @@ class CustomObtainAuthToken(ObtainAuthToken):
         # Lấy dữ liệu POST request
         email = request.data.get('email')
         password = request.data.get('password')
-        print(email + " " + password)
+        print(request.data)
         # Kiểm tra email và mật khẩu hợp lệ
         if email is None or password is None:
             return Response({'error': 'Please provide both username and password'},
@@ -322,18 +376,20 @@ class CustomObtainAuthToken(ObtainAuthToken):
         user = authenticate(email=email, password=password)
 
         if not user:
-            return Response({'error': 'Invalid Credentials'},
+            return Response({'error': 'Invalid Account'},
                             status=status.HTTP_404_NOT_FOUND)
         # Tạo hoặc lấy token xác thực
         token, created = Token.objects.get_or_create(user=user)
-        return Response({'token': token.key})
+        account = Account.objects.filter(email=email).first()
+        return Response({'token': token.key,
+                         "role": 1 if account.is_staff == True else 0})
     
 class LogoutView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
         request.user.auth_token.delete()
-        return Response(status=200)
+        return Response("Token is deleted",status=200)
 
           
 def get_home(request):
